+++ hlp/mt/config/config.json
@@ -5,28 +5,21 @@
   "en_tokenize_type": "BPE",
   "zh_tokenize_type": "CHAR",
   "path_to_train_file": "./data/corpus/en-zh_train.txt",
-  "path_to_train_file_en": "./data/corpus/train_en.txt",
-  "path_to_train_file_zh": "./data/corpus/train_zh.txt",
   "path_to_eval_file": "./data/corpus/en-zh_eval.txt",
-  "path_encoded_sequences_en": "./data/encoded_corpus/encoded_sequences_en.npy",
-  "path_encoded_sequences_zh": "./data/encoded_corpus/encoded_sequences_zh.npy",
+  "encoded_sequences_path_prefix": "./data/encoded_corpus/encoded_sequences_",
+  "tokenizer_path_prefix": "./data/tokenizer/tokenizer_",
   "num_eval": 5,
   "checkpoint_path_dir": "./checkpoints",
-  "checkpoint_path_en_zh": "./checkpoints/train_en_zh",
-  "checkpoint_path_zh_en": "./checkpoints/train_zh_en",
-  "en_bpe_tokenizer_path": "./data/tokenizer/en_tokenizer",
-  "en_tokenizer_path": "./data/tokenizer/en_tokenizer.json",
-  "zh_tokenizer_path": "./data/tokenizer/zh_tokenizer.json",
   "BUFFER_SIZE": 20000,
   "test_size": 0.001,
-  "num_sentences": 1000,
+  "num_sentences": 500,
   "BATCH_SIZE": 32,
   "num_layers": 4,
   "d_model": 256,
   "dff": 512,
   "num_heads": 8,
   "dropout_rate" : 0.1,
-  "EPOCHS": 2,
+  "EPOCHS": 50,
   "max_target_length": 200,
   "start_word": "<start>",
   "end_word": "<end>",
