+++ hlp/mt/nmt_train.py
@@ -2,8 +2,7 @@ import os
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
 from model import evaluate, trainer, translator
 from common import preprocess as _pre
-from model import network
-from model import transformer as _transformer
+from model import nmt_model
 
 """
 使用 ./data 文件夹下的指定文件(默认 en-ch.txt)进行训练
@@ -15,11 +14,10 @@ def main():
     vocab_size_source, vocab_size_target = _pre.train_preprocess()
 
     # 创建模型及相关变量
-    optimizer, train_loss, train_accuracy = _transformer.get_optimizer()
-    transformer = network.get_model(vocab_size_source, vocab_size_target)
+    transformer = nmt_model.get_model(vocab_size_source, vocab_size_target)
 
     # 开始训练
-    trainer.train(transformer, optimizer, train_loss, train_accuracy)
+    trainer.train(transformer)
 
 
 if __name__ == '__main__':
